update replication to use ReplicateLike
